Move encoding to public module
